adjust int data to double data

+++ 7주차/Data_structure_나윤재/나윤재_실습_07_1/NYJNode.h
@@ -1,15 +1,15 @@
 #include<cstdio>
 class Node{
 	Node* link;
-	int data;
+	double data;
 
 public:
 	Node(int val =0) : data(val),link(NULL){}
 	Node* getLink() { return link;}
 
 	void setLink(Node * next){link =next;}
-	void display(){printf(" <%2d>" ,data);}
-	bool hasData(int val) { return data == val;}
+	void display(){printf(" <%2lf>" ,data);}
+	bool hasData(double val) { return data == val;}
 
 	void insertNext(Node *n){
 		if(n!= NULL){
